fixed IoClassSerialiserTests for custom pojos that are directly serialised as rootObj
The corresponding routine in IoSerialiser
<E> E getCustomData(FieldSerialiser<E> serialiser);
does not provide any provisions to make in-place modification and thus always returns a new custom object. Thus the return value needs to be used for deserialisation.

N.B. this is different for custom objects that are nested within another (generic) pojo. In this case the regular return/read FieldSerialiser call-back functions are used.