Fix deserialisation of nested inner static classes
For fields with static inner classes as their type, the deserialiser
always assumes non-static inner classes and cannot find the constructor.
If one attempts to deserialise such an object the deserialiser will
throw `NoSuchMethodException` (which is caught and printed to the log
before continuing) when trying to allocate the field using the constructor
with the parent class as its default argument.

This commit adds a testcase for a nested class with static and
non-static nested class fields.
The fix consists of checking for the `static` modifier after determining
an inner class and treating static inner classes the same as normal
classes.