QueryParameterParser: prevent double encoding
The url constructor will encode the individual parts of the url in its
constructor. If the individual components are already percent encoded,
the percent signs of the encdoded characters will themself be encoded.

For example `http:://example.com/path/file.ext?type=foo/bar` would
get the query part encoded to `type=foo%2252Fbar` instead of
`type=foo/bar`. Note that which characters will be escaped in an url
depends on the part, in other components the slash would need to be
escaped. Fortunately this is already implemented in the uri
implementation.

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>